% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-flatten.R
\name{flatten_epi_columns}
\alias{flatten_epi_columns}
\title{Flatten nested extraction columns with dot notation}
\usage{
flatten_epi_columns(data, names_sep = ".")
}
\arguments{
\item{data}{A data frame containing nested objects or df-cols.}

\item{names_sep}{Separator used between parent and child names.}
}
\value{
A flat tibble.
}
\description{
Flatten nested extraction columns with dot notation
}
