% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{partial_date_errors}
\alias{partial_date_errors}
\title{Retrieve partial-date reconstruction errors}
\usage{
partial_date_errors(x)
}
\arguments{
\item{x}{An object returned by \code{\link[=rebuild_epi_partial_dates]{rebuild_epi_partial_dates()}}.}
}
\value{
A tibble with columns \code{row}, \code{column}, \code{value}, and \code{error}.
}
\description{
Retrieve partial-date reconstruction errors
}
