% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess_errors}
\alias{postprocess_errors}
\title{Retrieve post-processing errors from extracted data}
\usage{
postprocess_errors(x)
}
\arguments{
\item{x}{An object returned by \code{\link[=extract_epi_data]{extract_epi_data()}}.}
}
\value{
A tibble with columns \code{row}, \code{column}, \code{value}, and \code{error}.
}
\description{
Retrieve post-processing errors from extracted data
}
