% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_case}
\alias{type_epi_case}
\title{Case-level epidemiological record}
\usage{
type_epi_case(required = FALSE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Case-level epidemiological record
}
