% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_case_report}
\alias{type_epi_case_report}
\title{Report-level extraction schema}
\usage{
type_epi_case_report(required = TRUE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Report-level extraction schema
}
