% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_case_status}
\alias{type_epi_case_status}
\title{Epidemiology-specific case status field}
\usage{
type_epi_case_status(
  description = "Case status classification.",
  required = FALSE
)
}
\arguments{
\item{description}{Field description supplied to the model.}

\item{required}{Should the field be required?}
}
\value{
An \code{ellmer} enum type.
}
\description{
Epidemiology-specific case status field
}
