% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_contact}
\alias{type_epi_contact}
\title{Contact associated with a case}
\usage{
type_epi_contact(required = FALSE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Contact associated with a case
}
