% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_counts}
\alias{type_epi_counts}
\title{Aggregate epidemiological counts}
\usage{
type_epi_counts(required = FALSE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Aggregate epidemiological counts
}
