% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_demographics}
\alias{type_epi_demographics}
\title{Demographic attributes for an individual or group}
\usage{
type_epi_demographics(required = FALSE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Demographic attributes for an individual or group
}
