% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_event}
\alias{type_epi_event}
\title{Event-level epidemiological attributes}
\usage{
type_epi_event(required = FALSE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Event-level epidemiological attributes
}
