% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_exposure}
\alias{type_epi_exposure}
\title{Exposure event relevant to a case interview}
\usage{
type_epi_exposure(required = FALSE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Exposure event relevant to a case interview
}
