% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_illness}
\alias{type_epi_illness}
\title{Symptom and illness details for an individual case}
\usage{
type_epi_illness(required = FALSE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Symptom and illness details for an individual case
}
