% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_interview}
\alias{type_epi_interview}
\title{Case interview metadata}
\usage{
type_epi_interview(required = FALSE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Case interview metadata
}
