% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_location}
\alias{type_epi_location}
\title{Geographic location type for epidemiological extraction}
\usage{
type_epi_location(required = FALSE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Geographic location type for epidemiological extraction
}
