% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_notification}
\alias{type_epi_notification}
\title{Laboratory and notification details for a case}
\usage{
type_epi_notification(required = FALSE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Laboratory and notification details for a case
}
