% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_outbreak}
\alias{type_epi_outbreak}
\title{Outbreak-level extraction schema}
\usage{
type_epi_outbreak(required = TRUE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Outbreak-level extraction schema
}
