% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_outcome}
\alias{type_epi_outcome}
\title{Epidemiology-specific outcome field}
\usage{
type_epi_outcome(description = "Clinical outcome.", required = FALSE)
}
\arguments{
\item{description}{Field description supplied to the model.}

\item{required}{Should the field be required?}
}
\value{
An \code{ellmer} enum type.
}
\description{
Epidemiology-specific outcome field
}
