% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_person}
\alias{type_epi_person}
\title{Person-level epidemiological attributes}
\usage{
type_epi_person(required = FALSE)
}
\arguments{
\item{required}{Should the object be required?}
}
\value{
An \code{ellmer} object type.
}
\description{
Person-level epidemiological attributes
}
