% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_test_result}
\alias{type_epi_test_result}
\title{Epidemiology-specific test result field}
\usage{
type_epi_test_result(description = "Laboratory test result.", required = FALSE)
}
\arguments{
\item{description}{Field description supplied to the model.}

\item{required}{Should the field be required?}
}
\value{
An \code{ellmer} enum type.
}
\description{
Epidemiology-specific test result field
}
