% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_epi_yes_no_unknown}
\alias{type_epi_yes_no_unknown}
\title{Epidemiology-specific tri-state field}
\usage{
type_epi_yes_no_unknown(
  description = "One of yes, no, or unknown.",
  required = FALSE
)
}
\arguments{
\item{description}{Field description supplied to the model.}

\item{required}{Should the field be required?}
}
\value{
An \code{ellmer} enum type.
}
\description{
Epidemiology-specific tri-state field
}
